BYOP Non-Content Update
I know, it's a strange way of putting it.. Suffice it to say, though, that the BYOP page is now currently set up not only to pull everything in from the database, but to pass the pizza along to the cart (and redirect there) when a valid pizza (Currently just having crust and size, not going into the logic of checking if it's a valid combo yet - Would be best to leave that out for the moment) has been submitted. Next would probably be getting the navbar and the like in place (Along with a bit of 'Sure you want to leave?' from Javascript perhaps), updating the styles to use the CSS, and.. Well, that's about it, I'd think. Need to test this when in class and all. Also, need to make the placeholder 'cart.php' page.