Lotsa stuff!
Namely, an updated SQL thing for building the database with a few sample pizza-related things in it, an updated BYOP and CartItem that not only work, but aren't spouting off errors left right and center. Added a (VERY) placeholder cart page, need to add some way of going back to another page from there. Also added a LuigiPizzaDB_Connect file for easy drop-in database connection (Goes by the name 'dbMaster'), and.. Well, not a whole lot else. Everything seems to be working, so I'm happy with the progress so far today. ^-^