Merging edits back into the main branch
Because, honestly, not entirely sure what changed.

Oh, also - Added another failsafe to pizza adding, a way to remove items from the cart, and a way to return to the BYOP page from the cart. Yay.


Now all I need is to plug in the nav and whatever CSS stuff is being used for this. Given current scope, I'd say the pizza builder is fully ar.. Functional. Let's go with functional instead of an overused reference.